refactor: extract mock_ableton_connection fixture to conftest.py
Move duplicated fixture from test_arrangement_clips.py,
test_recording.py, test_arrangement_markers.py, and
test_arrangement_navigation.py into shared tests/conftest.py.
