+++ tests/conftest.py
@@ -3,9 +3,19 @@
 import json
 import socket
 import threading
-from unittest.mock import patch
+from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
+
+
+@pytest.fixture
+def mock_ableton_connection():
+    """Create a mock Ableton connection for MCP server tests."""
+    with patch("MCP_Server.server.get_ableton_connection") as mock_get_conn:
+        mock_conn = MagicMock()
+        mock_conn.send_command_async = AsyncMock()
+        mock_get_conn.return_value = mock_conn
+        yield mock_conn
 
 
 class MockSocket:
